requirements.txt
# General packages
pyyaml
kaggle
prompter

# ML packages
pandas
scikit-learn
tensorboardX
optuna

# Jupyter packages
jupyterlab
seaborn
tqdm
ipywidgets

# NLP modules
nltk
torchtext
gensim
spacy
sentencepiece
